+++ mobile/android/android-components/components/feature/downloads/src/main/java/mozilla/components/feature/downloads/DownloadsFeature.kt
@@ -4,27 +4,31 @@
 
 package mozilla.components.feature.downloads
 
-import android.annotation.SuppressLint
 import android.content.Context
 import android.widget.Toast
 import androidx.annotation.VisibleForTesting
 import androidx.annotation.VisibleForTesting.PRIVATE
 import androidx.fragment.app.FragmentManager
-import mozilla.components.browser.session.Download
-import mozilla.components.browser.session.SelectionAwareSessionObserver
-import mozilla.components.browser.session.Session
-import mozilla.components.browser.session.SessionManager
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.cancel
+import kotlinx.coroutines.flow.collect
+import kotlinx.coroutines.flow.mapNotNull
+import mozilla.components.browser.state.selector.findCustomTabOrSelectedTab
+import mozilla.components.browser.state.state.SessionState
+import mozilla.components.browser.state.state.content.DownloadState
+import mozilla.components.browser.state.store.BrowserStore
 import mozilla.components.feature.downloads.DownloadDialogFragment.Companion.FRAGMENT_TAG
 import mozilla.components.feature.downloads.manager.AndroidDownloadManager
 import mozilla.components.feature.downloads.manager.DownloadManager
 import mozilla.components.feature.downloads.manager.OnDownloadCompleted
 import mozilla.components.feature.downloads.manager.noop
+import mozilla.components.lib.state.ext.flowScoped
 import mozilla.components.support.base.feature.LifecycleAwareFeature
 import mozilla.components.support.base.feature.OnNeedToRequestPermissions
 import mozilla.components.support.base.feature.PermissionsFeature
-import mozilla.components.support.base.observer.Consumable
 import mozilla.components.support.ktx.android.content.appName
 import mozilla.components.support.ktx.android.content.isPermissionGranted
+import mozilla.components.support.ktx.kotlinx.coroutines.flow.ifChanged
 
 /**
  * Feature implementation to provide download functionality for the selected
@@ -38,7 +42,8 @@ import mozilla.components.support.ktx.android.content.isPermissionGranted
  * @property onDownloadCompleted a callback invoked when a download is completed.
  * @property downloadManager a reference to the [DownloadManager] which is
  * responsible for performing the downloads.
- * @property sessionManager a reference to the application's [SessionManager].
+ * @property store a reference to the application's [BrowserStore].
+ * @property useCases [DownloadsUseCases] instance for consuming processed downloads.
  * @property fragmentManager a reference to a [FragmentManager]. If a fragment
  * manager is provided, a dialog will be shown before every download.
  * @property dialog a reference to a [DownloadDialogFragment]. If not provided, an
@@ -46,15 +51,16 @@ import mozilla.components.support.ktx.android.content.isPermissionGranted
  */
 class DownloadsFeature(
     private val applicationContext: Context,
+    private val store: BrowserStore,
+    private val useCases: DownloadsUseCases,
     override var onNeedToRequestPermissions: OnNeedToRequestPermissions = { },
     onDownloadCompleted: OnDownloadCompleted = noop,
     private val downloadManager: DownloadManager = AndroidDownloadManager(applicationContext),
-    sessionManager: SessionManager,
-    private val sessionId: String? = null,
+    private val customTabId: String? = null,
     private val fragmentManager: FragmentManager? = null,
     @VisibleForTesting(otherwise = PRIVATE)
     internal var dialog: DownloadDialogFragment = SimpleDownloadDialogFragment.newInstance()
-) : SelectionAwareSessionObserver(sessionManager), LifecycleAwareFeature, PermissionsFeature {
+) : LifecycleAwareFeature, PermissionsFeature {
 
     var onDownloadCompleted: OnDownloadCompleted
         get() = downloadManager.onDownloadCompleted
@@ -64,15 +70,26 @@ class DownloadsFeature(
         this.onDownloadCompleted = onDownloadCompleted
     }
 
+    private var scope: CoroutineScope? = null
+
     /**
      * Starts observing downloads on the selected session and sends them to the [DownloadManager]
      * to be processed.
      */
     override fun start() {
-        observeIdOrSelected(sessionId)
-
         findPreviousDialogFragment()?.let {
-            reAttachOnStartDownloadListener(it)
+            dialog = it
+        }
+
+        scope = store.flowScoped { flow ->
+            flow.mapNotNull { state -> state.findCustomTabOrSelectedTab(customTabId) }
+                .ifChanged { it.content.download }
+                .collect { state ->
+                    val download = state.content.download
+                    if (download != null) {
+                        processDownload(state, download)
+                    }
+                }
         }
     }
 
@@ -80,18 +97,18 @@ class DownloadsFeature(
      * Stops observing downloads on the selected session.
      */
     override fun stop() {
-        super.stop()
+        scope?.cancel()
+
         downloadManager.unregisterListeners()
     }
 
     /**
      * Notifies the [DownloadManager] that a new download must be processed.
      */
-    @SuppressLint("MissingPermission")
-    override fun onDownload(session: Session, download: Download): Boolean {
+    private fun processDownload(tab: SessionState, download: DownloadState): Boolean {
         return if (applicationContext.isPermissionGranted(downloadManager.permissions.asIterable())) {
             if (fragmentManager != null) {
-                showDialog(download, session)
+                showDialog(tab, download)
                 false
             } else {
                 startDownload(download)
@@ -102,12 +119,12 @@ class DownloadsFeature(
         }
     }
 
-    private fun startDownload(download: Download): Boolean {
+    private fun startDownload(download: DownloadState): Boolean {
         val id = downloadManager.download(download)
         return if (id != null) {
             true
         } else {
-            showUnSupportFileErrorMessage()
+            showDownloadNotSupportedError()
             false
         }
     }
@@ -117,30 +134,44 @@ class DownloadsFeature(
      * either trigger or clear the pending download.
      */
     override fun onPermissionsResult(permissions: Array<String>, grantResults: IntArray) {
-        if (applicationContext.isPermissionGranted(downloadManager.permissions.asIterable())) {
-            activeSession?.let { session ->
-                session.download.consume { onDownload(session, it) }
+        if (permissions.isEmpty()) {
+            // If we are requesting permissions while a permission prompt is already being displayed
+            // then Android seems to call `onPermissionsResult` immediately with an empty permissions
+            // list. In this case just ignore it.
+            return
+        }
+
+        withActiveDownload { (tab, download) ->
+            if (applicationContext.isPermissionGranted(downloadManager.permissions.asIterable())) {
+                processDownload(tab, download)
+            } else {
+                useCases.consumeDownload(tab.id, download.id)
             }
-        } else {
-            activeSession?.download = Consumable.empty()
         }
     }
 
-    private fun showUnSupportFileErrorMessage() {
-        val text = applicationContext.getString(
-            R.string.mozac_feature_downloads_file_not_supported2,
-            applicationContext.appName
-        )
-
-        Toast.makeText(applicationContext, text, Toast.LENGTH_LONG).show()
+    @VisibleForTesting(otherwise = PRIVATE)
+    internal fun showDownloadNotSupportedError() {
+        Toast.makeText(
+            applicationContext,
+            applicationContext.getString(
+                R.string.mozac_feature_downloads_file_not_supported2,
+                applicationContext.appName),
+                Toast.LENGTH_LONG
+        ).show()
     }
 
-    @SuppressLint("MissingPermission")
-    private fun showDialog(download: Download, session: Session) {
+    private fun showDialog(tab: SessionState, download: DownloadState) {
         dialog.setDownload(download)
 
         dialog.onStartDownload = {
-            session.download.consume(this::startDownload)
+            startDownload(download)
+
+            useCases.consumeDownload.invoke(tab.id, download.id)
+        }
+
+        dialog.onCancelDownload = {
+            useCases.consumeDownload.invoke(tab.id, download.id)
         }
 
         if (!isAlreadyADialogCreated() && fragmentManager != null) {
@@ -152,16 +183,13 @@ class DownloadsFeature(
         return findPreviousDialogFragment() != null
     }
 
-    private fun reAttachOnStartDownloadListener(previousDialog: DownloadDialogFragment?) {
-        previousDialog?.let {
-            dialog = it
-            activeSession?.let { session ->
-                session.download.consume { download -> onDownload(session, download) }
-            }
-        }
-    }
-
     private fun findPreviousDialogFragment(): DownloadDialogFragment? {
         return fragmentManager?.findFragmentByTag(FRAGMENT_TAG) as? DownloadDialogFragment
+    }
+
+    private fun withActiveDownload(block: (Pair<SessionState, DownloadState>) -> Unit) {
+        val state = store.state.findCustomTabOrSelectedTab(customTabId) ?: return
+        val download = state.content.download ?: return
+        block(Pair(state, download))
     }
 }